Enclose color escape sequences in \x01 and \x02
If the invisible ANSI escape sequences for colors are not wrapped in the
characters "\x01" and "\x02", the escape sequences count towards the
length of the PS1 prompt. As a result, the terminal wraps the command
being typed to the next line much before the end of the line is actually
reached.

Therefore enclose the invisible escape sequences in "\x01" and "\x02" so
that the escape sequences do not count towards the length of the PS1
prompt.